All Done

+++ src/main/kotlin/com/springBoot/note_app/controllers/NoteController.kt
@@ -3,7 +3,11 @@ package com.springBoot.note_app.controllers
 import com.springBoot.note_app.controllers.NoteController.NoteResponse
 import com.springBoot.note_app.database.model.Note
 import com.springBoot.note_app.database.repository.NoteRepository
+import jakarta.validation.Valid
+import jakarta.validation.constraints.NotBlank
 import org.bson.types.ObjectId
+import org.springframework.security.core.context.SecurityContext
+import org.springframework.security.core.context.SecurityContextHolder
 import org.springframework.web.bind.annotation.*
 import java.time.Instant
 
@@ -15,6 +19,7 @@ class NoteController(
 
     data class NoteRequest(
         val id : String?,
+        @field:NotBlank(message = "Title can't be blank.")
         val title : String,
         val content : String,
         val color : Long
@@ -30,8 +35,9 @@ class NoteController(
 
     @PostMapping
     fun save(
-        @RequestBody noteRequest: NoteRequest
+        @Valid @RequestBody noteRequest: NoteRequest
     ) : NoteResponse{
+        val ownerId = SecurityContextHolder.getContext().authentication.principal as String
         val note = repository.save(
             Note(
                 id = noteRequest.id?.let { ObjectId(it) } ?: ObjectId.get(),
@@ -39,16 +45,15 @@ class NoteController(
                 content = noteRequest.content,
                 color = noteRequest.color,
                 createdAt = Instant.now(),
-                ownerId = ObjectId()
+                ownerId = ObjectId(ownerId)
             )
         )
         return note.toNoteResponse()
     }
 
     @GetMapping
-    fun findByOwnerId(
-        @RequestParam(required = true) ownerId: String
-    ): List<NoteResponse> {
+    fun findByOwnerId(): List<NoteResponse> {
+        val ownerId = SecurityContextHolder.getContext().authentication.principal as String
         return repository.findByOwnerId(ObjectId(ownerId)).map {
             it.toNoteResponse()
         }
@@ -56,7 +61,13 @@ class NoteController(
 
     @DeleteMapping(path = ["/{id}"])
     fun deleteById(@PathVariable id:String){
-        repository.deleteById(ObjectId(id))
+        val note = repository.findById(ObjectId(id)).orElseThrow{
+            IllegalArgumentException("Note not found")
+        }
+        val ownerId = SecurityContextHolder.getContext().authentication.principal as String
+        if(note.ownerId.toHexString() == ownerId) {
+            repository.deleteById(ObjectId(id))
+        }
     }
 
 }
